% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opportunity.R
\docType{data}
\name{opportunity}
\alias{opportunity}
\title{Chetty et al data on intergenerational mobility across US Census Zones.}
\format{
An object of class \code{"cross"}.
}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(opportunity)
}
\description{
A subset of their variables with 709 observations containing information on the following variables.
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
