% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_basis.R
\name{optimal_basis}
\alias{optimal_basis}
\title{Plot BIC and R2 for different spatial bases}
\usage{
optimal_basis(fm, df, max_splines, Description = "")
}
\arguments{
\item{fm}{The regression formula.}

\item{df}{The dataset used.}

\item{max_splines}{The maximum dimension of tensor to examine}

\item{Description}{Optional description of the regression used to label the output diagram.}
}
\value{
ggplot object: graph of BIC and R2 of regression of dependent variable on bases.
}
\description{
\code{optimal_basis()} gives a plot of BIC and adjusted R-squared against number of
principal components for tensor bases starting at 3x3. It reports the combination
that minimizes BIC.
}
\examples{
library(spatInfer)
data(opportunity)
#Choose a sample of 100 to speed things up.
set.seed(123)
opportunity=opportunity |> dplyr::slice_sample(n=100)
optimal_basis(mobility~racial_seg+single_mom,opportunity,
    max_splines=6, Description="Intergenerational Mobility.")
}
