% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo_table.R
\name{placebo_table}
\alias{placebo_table}
\title{Makes a summary table of estimated and placebo p values from output generated by BHC or IM placebos.}
\usage{
placebo_table(plbo, adjust = "BCH")
}
\arguments{
\item{plbo}{An object created by placebo_bch or placebo_im.}

\item{adjust}{Label the table with BCH or IM according to which one is chosen.}
}
\value{
Generates a tinytable object. This gives the regression and placebo p values and the proportion
of placebo simulation significant at 5\\% as the number of clusters increases.
The regression confidence interval and its width are also reported. The next part of the table gives
the z-score of a Moran test, using 5 nearest neighbours, the R2 of a regression of the treatment on the
principal components of the tensor, and the effective range and structure of this regression's residuals.
}
\description{
Makes a summary table of estimated and placebo p values from output generated by BHC or IM placebos.
}
\examples{
library(spatInfer)
data(opportunity)
# Choose a sample of 250 observations and run only nSim=100 placebo simulations to speed things up.
set.seed(123)
opportunity=opportunity |> dplyr::slice_sample(n=250)
# Use the number of splines and PCs indicated by optimal_basis()
plbo_bch=placebo_bch(mobility~racial_seg+single_mom,  opportunity,
splines=4, pc_num=3,nSim=100)

placebo_table(plbo_bch,adjust="BCH")
}
